// reference/return-requests/domain/return-request-status.enum.ts
export enum ReturnRequestStatusEnum {
  PENDING = 'pending',
  APPROVED = 'approved',
  REJECTED = 'rejected',
  PICKUP_SCHEDULED = 'pickup_scheduled',
  PICKED_UP = 'picked_up',
  RECEIVED = 'received',
  REFUND_PROCESSING = 'refund_processing',
  REFUNDED = 'refunded',
  REFUND_FAILED = 'refund_failed',
  CLOSED = 'closed',
}